refactor: (wip) change almost everything
