+++ src/Storageless/Service/StoragelessSession.php
@@ -20,154 +20,148 @@ declare(strict_types=1);
 
 namespace PSR7Sessions\Storageless\Service;
 
+use BadMethodCallException;
 use DateInterval;
-use DateTimeImmutable;
-use Dflydev\FigCookies\Cookie;
-use Dflydev\FigCookies\FigRequestCookies;
 use Dflydev\FigCookies\FigResponseCookies;
-use Dflydev\FigCookies\Modifier\SameSite;
 use Dflydev\FigCookies\SetCookie;
-use Dflydev\FigCookies\SetCookies;
-use Exception;
-use Lcobucci\Clock\SystemClock;
-use Lcobucci\JWT\Configuration;
+use InvalidArgumentException;
 use Lcobucci\JWT\Encoding\ChainedFormatter;
-use Lcobucci\JWT\Signer;
+use Lcobucci\JWT\Token;
 use Lcobucci\JWT\UnencryptedToken;
 use Lcobucci\JWT\Validation\Constraint\SignedWith;
 use Lcobucci\JWT\Validation\Constraint\StrictValidAt;
-use Psr\Clock\ClockInterface;
-use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Message\ServerRequestInterface;
-use PSR7Sessions\Storageless\Session\DefaultSessionData;
+use Psr\Http\Server\RequestHandlerInterface;
+use PSR7Sessions\Storageless\Http\ClientFingerprint\SameOriginRequest;
+use PSR7Sessions\Storageless\Http\Configuration;
+use PSR7Sessions\Storageless\Http\SessionMiddleware;
+use PSR7Sessions\Storageless\Session\LazySession;
 use PSR7Sessions\Storageless\Session\SessionInterface;
-use stdClass;
-use Throwable;
 
-use function array_key_exists;
-use function assert;
-use function is_string;
 use function sprintf;
 
 final class StoragelessSession implements SessionStorage
 {
-    private const DEFAULT_COOKIE       = '__Secure-slsession';
-    private const SESSION_CLAIM        = 'session-data';
-    private const DEFAULT_IDLE_TIMEOUT = 300;
-
-    private readonly Cookie|SetCookie $cookie;
+    public const SESSION_CLAIM = 'session-data';
 
     public function __construct(
-        private readonly Configuration $configuration,
-        private readonly int $idleTimeout,
-        Cookie|SetCookie $cookie,
-        private readonly ClockInterface $clock,
+        private readonly Configuration $config,
     ) {
-        $this->cookie = clone $cookie;
     }
 
-    public static function fromSymmetricKeyDefaults(
-        Signer\Key $symmetricKey,
-        int $idleTimeout = self::DEFAULT_IDLE_TIMEOUT,
-        Cookie|SetCookie|null $cookie = null,
-        ClockInterface|null $clock = null,
-    ): self {
-        return new self(
-            Configuration::forSymmetricSigner(
-                new Signer\Hmac\Sha256(),
-                $symmetricKey,
-            ),
-            $idleTimeout,
-            $cookie ?? SetCookie::create(self::DEFAULT_COOKIE)
-                ->withSecure(true)
-                ->withHttpOnly(true)
-                ->withSameSite(SameSite::lax())
-                ->withPath('/'),
-            $clock ?? SystemClock::fromUTC(),
-        );
-    }
-
-    public static function fromRsaAsymmetricKeyDefaults(
-        Signer\Key $privateRsaKey,
-        Signer\Key $publicRsaKey,
-        int $idleTimeout = self::DEFAULT_IDLE_TIMEOUT,
-        Cookie|SetCookie|null $cookie = null,
-        ClockInterface|null $clock = null,
-    ): self {
-        return new self(
-            Configuration::forAsymmetricSigner(
-                new Signer\Rsa\Sha256(),
-                $privateRsaKey,
-                $publicRsaKey,
-            ),
-            $idleTimeout,
-            $cookie ?? SetCookie::create(self::DEFAULT_COOKIE)
-                ->withSecure(true)
-                ->withHttpOnly(true)
-                ->withSameSite(SameSite::lax())
-                ->withPath('/'),
-            $clock ?? SystemClock::fromUTC(),
-        );
-    }
-
-    public function withSession(ServerRequestInterface|ResponseInterface $message, SessionInterface $session): RequestInterface|ResponseInterface
+    public function appendSession(SessionInterface $session, ServerRequestInterface $request, Response|null $response = null, RequestHandlerInterface|null $handler = null): Response
     {
-        if ($message instanceof ResponseInterface) {
-            return $this->withResponseSession($message, $session, $this->clock->now());
-        }
+        $sameOriginRequest = new SameOriginRequest($this->config->getClientFingerprintConfiguration(), $request);
+        $token             = $this->requestToToken($request, $sameOriginRequest);
 
-        return $this->withRequestSession($message, $session, $this->clock->now());
-    }
+        $middleware = new SessionMiddleware($this->config);
 
-    public function get(ServerRequestInterface|ResponseInterface $message): SessionInterface
-    {
-        $cookie = $this->getCookieFromMessage($message);
-
-        return $cookie === null
-            ? DefaultSessionData::newEmptySession()
-            : $this->cookieToSession($cookie);
-    }
-
-    public function getCookieFromMessage(ServerRequestInterface|ResponseInterface $message): SetCookie|Cookie|null
-    {
-        // TODO: Why we cannot use Cookies::fromRequest() ?
-        // See: https://github.com/dflydev/dflydev-fig-cookies/issues/57
-        if ($message instanceof ServerRequestInterface) {
-            $cookies = $message->getCookieParams();
-
-            if (! array_key_exists($this->cookie->getName(), $cookies)) {
-                return null;
+        /** @var RequestHandlerInterface $handler */
+        $handler ??= fn (ResponseInterface $response): RequestHandlerInterface => new class ($response) implements RequestHandlerInterface {
+            public function __construct(
+                private readonly ResponseInterface $response,
+            ) {
             }
 
-            $cookieValue = $cookies[$this->cookie->getName()];
-            assert(is_string($cookieValue));
+            public function handle(Request $request): ResponseInterface
+            {
+              return $this->response;
+            }
+        };
 
-            return Cookie::create($this->cookie->getName(), $cookieValue === '' ? null : $cookieValue);
-        }
+        $response ??= $handler->handle($request->withAttribute($this->config->getSessionAttribute(), $session));
 
-        return SetCookies::fromResponse($message)->get($this->cookie->getName());
+        return $this->appendToken(
+            $session,
+            $response,
+            $token,
+            $sameOriginRequest,
+        );
+
+        return $middleware->process($request, $handler($response));
     }
 
-    public function cookieToToken(SetCookie|Cookie|null $cookie): UnencryptedToken|null
+    public function getSession(Request $request): SessionInterface
     {
-        if ($cookie === null) {
+        $sameOriginRequest = new SameOriginRequest($this->config->getClientFingerprintConfiguration(), $request);
+        $token             = $this->requestToToken($request, $sameOriginRequest);
+
+        return LazySession::fromToken($token);
+    }
+
+    /**
+     * @throws BadMethodCallException
+     * @throws InvalidArgumentException
+     */
+    private function appendToken(
+        SessionInterface $sessionContainer,
+        Response $response,
+        Token|null $token,
+        SameOriginRequest $sameOriginRequest,
+    ): Response {
+        $sessionContainerChanged = $sessionContainer->hasChanged();
+
+        if ($sessionContainerChanged && $sessionContainer->isEmpty()) {
+            return FigResponseCookies::set($response, $this->getExpirationCookie());
+        }
+
+        if ($sessionContainerChanged || $this->shouldTokenBeRefreshed($token)) {
+            return FigResponseCookies::set($response, $this->getTokenCookie($sessionContainer, $sameOriginRequest));
+        }
+
+        return $response;
+    }
+
+    /** @throws BadMethodCallException */
+    private function getTokenCookie(SessionInterface $sessionContainer, SameOriginRequest $sameOriginRequest): SetCookie
+    {
+        $now       = $this->config->getClock()->now();
+        $expiresAt = $now->add(new DateInterval(sprintf('PT%sS', $this->config->getIdleTimeout())));
+
+        $jwtConfiguration = $this->config->getJwtConfiguration();
+
+        $builder = $jwtConfiguration->builder(ChainedFormatter::withUnixTimestampDates())
+            ->issuedAt($now)
+            ->canOnlyBeUsedAfter($now)
+            ->expiresAt($expiresAt)
+            ->withClaim(self::SESSION_CLAIM, $sessionContainer);
+
+        $builder = $sameOriginRequest->configure($builder);
+
+        return $this
+            ->config->getCookie()
+            ->withValue(
+                $builder
+                    ->getToken($jwtConfiguration->signer(), $jwtConfiguration->signingKey())
+                    ->toString(),
+            )
+            ->withExpires($expiresAt);
+    }
+
+    public function requestToToken(Request $request, SameOriginRequest|null $sameOriginRequest = null): UnencryptedToken|null
+    {
+        $sameOriginRequest ??= new SameOriginRequest($this->config->getClientFingerprintConfiguration(), $request);
+
+        /** @var array<string, string> $cookies */
+        $cookies    = $request->getCookieParams();
+        $cookieName = $this->config->getCookie()->getName();
+
+        if (! isset($cookies[$cookieName])) {
             return null;
         }
 
-        $jwt = $cookie->getValue();
-
-        if ($jwt === null) {
+        $cookie = $cookies[$cookieName];
+        if ($cookie === '') {
             return null;
         }
 
-        if ($jwt === '') {
-            return null;
-        }
-
+        $jwtConfiguration = $this->config->getJwtConfiguration();
         try {
-            $token = $this->configuration->parser()->parse($jwt);
-        } catch (Throwable) {
+            $token = $jwtConfiguration->parser()->parse($cookie);
+        } catch (InvalidArgumentException) {
             return null;
         }
 
@@ -175,92 +169,41 @@ final class StoragelessSession implements SessionStorage
             return null;
         }
 
-        $isValid = $this
-            ->configuration
-            ->validator()
-            ->validate(
-                $token,
-                new StrictValidAt($this->clock),
-                new SignedWith($this->configuration->signer(), $this->configuration->verificationKey()),
-            );
+        $constraints = [
+            new StrictValidAt($this->config->getClock()),
+            new SignedWith($jwtConfiguration->signer(), $jwtConfiguration->verificationKey()),
+            $sameOriginRequest,
+        ];
 
-        if ($isValid === false) {
+        if (! $jwtConfiguration->validator()->validate($token, ...$constraints)) {
             return null;
         }
 
         return $token;
     }
 
-    private function withRequestSession(RequestInterface $request, SessionInterface $session, DateTimeImmutable $now): RequestInterface
+    private function getExpirationCookie(): SetCookie
     {
-        if ($session->hasChanged() === false) {
-            return $request;
-        }
-
-        if (! $this->cookie instanceof Cookie) {
-            throw new Exception(
-                'The default cookie is not a Cookie type.',
+        return $this
+            ->config->getCookie()
+            ->withValue(null)
+            ->withExpires(
+                $this->config->getClock()
+                    ->now()
+                    ->modify('-30 days'),
             );
-        }
-
-        return FigRequestCookies::set(
-            $request,
-            $this->appendCookieSession($this->cookie, $session, $now),
-        );
     }
 
-    private function withResponseSession(ResponseInterface $response, SessionInterface $session, DateTimeImmutable $now): ResponseInterface
+    private function shouldTokenBeRefreshed(Token|null $token): bool
     {
-        if (! $this->cookie instanceof SetCookie) {
-            throw new Exception(
-                'The default cookie is not a SetCookie type.',
-            );
-        }
-
-        if ($session->isEmpty()) {
-            return FigResponseCookies::set(
-                $response,
-                $this->cookie->withExpires($now->modify('-30 days')),
-            );
-        }
-
-        return FigResponseCookies::set(
-            $response,
-            $this
-                ->appendCookieSession(
-                    $this->cookie->withExpires($now->add(new DateInterval(sprintf('PT%sS', $this->idleTimeout)))),
-                    $session,
-                    $now,
-                ),
-        );
-    }
-
-    /** @psalm-return ($cookie is SetCookie ? SetCookie : Cookie) */
-    private function appendCookieSession(SetCookie|Cookie $cookie, SessionInterface $session, DateTimeImmutable $now): SetCookie|Cookie
-    {
-        $value = $session->isEmpty()
-            ? null
-            : $this->configuration->builder(ChainedFormatter::withUnixTimestampDates())
-                ->issuedAt($now)
-                ->canOnlyBeUsedAfter($now)
-                ->expiresAt($now->add(new DateInterval(sprintf('PT%sS', $this->idleTimeout))))
-                ->withClaim(self::SESSION_CLAIM, $session)
-                ->getToken($this->configuration->signer(), $this->configuration->signingKey())
-                ->toString();
-
-        return $cookie->withValue($value);
-    }
-
-    private function cookieToSession(SetCookie|Cookie $cookie): SessionInterface
-    {
-        $token = $this->cookieToToken($cookie);
-
         if ($token === null) {
-            return DefaultSessionData::newEmptySession();
+            return false;
         }
 
-        return DefaultSessionData::fromDecodedTokenData(
-            (object) $token->claims()->get(self::SESSION_CLAIM, new stdClass()),
+        return $token->hasBeenIssuedBefore(
+            $this->config->getClock()
+                ->now()
+                ->sub(new DateInterval(sprintf('PT%sS', $this->config->getRefreshTime()))),
         );
     }
 }
